debt - move pick and open into electron main service
